+++ app/Http/Controllers/MilkProductController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 
 use App\Models\MilkProduct;
 use App\Models\ingredient;
-
+use Illuminate\Support\Facades\Auth;
 
 class MilkProductController extends Controller
 {
@@ -14,6 +14,11 @@ class MilkProductController extends Controller
 
     public function index()
     {
+         if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
+
         $milkProducts=MilkProduct::with('ingredients')->get();
 
         return view('milk_products.index',['milkProducts'=>$milkProducts]);
@@ -21,12 +26,20 @@ class MilkProductController extends Controller
 
     public function create()
     {
+        if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
+
         return view('milk_products.create');
     }
 
     public function store(Request $request)
     {
-
+        if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
   
   
         $request->validate([
@@ -70,13 +83,23 @@ class MilkProductController extends Controller
 
     public function edit(Request $request,MilkProduct $milkProduct)
     {
-        
+        if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
+
         return view('milk_products.edit',['milkProduct'=>$milkProduct]);
 
     }
 
     public function update(Request $request, MilkProduct $milkProduct)
-{
+    {
+
+        if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
+
     $data = $request->validate([
         'product_name' => 'required|string|max:255',
         'unit_price' => 'required|numeric|min:1',
@@ -102,16 +125,26 @@ class MilkProductController extends Controller
     }
 
     return redirect()->route('milk_product.list')->with('success', 'Milk product updated successfully!');
-}
+    }
 
 
     public function view(MilkProduct $milkProduct)
     {
+        if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
+
         return view('milk_products.view',['milkProduct'=>$milkProduct]);
     }
 
     public function destroy(MilkProduct $milkProduct)
     {
+        if (!in_array(Auth::user()->role_id, [1, 6, 5])) 
+        {
+            abort(403, 'Unauthorized action.');
+        }
+
         $milkProduct->delete();
         return redirect()->route('milk_product.list');
     }
